move social force to examples write docstring for abstract base
